go-bindings: Extend the lifetime of C allocated objects
This is to make sure the garbage collector doesn't call the finalizer of those objects (i.e. deallocate the related C-memory) while they are still used in C parts of the bindings. If a object `obj` is delivered as parameter to the function `F()` and there is no other use after the call of `F()` the GC seems to be allowed to free it up / finalize it (See https://golang.org/pkg/runtime/#KeepAlive). The use of pointer types as caller types is required otherwise it doesn't not call it on the actual object but rather a copy it seems since we use pointer for all allocated objects now..