Don't call core_set_thread_initializer if not defined
`core_set_thread_initializer` is only defined if macro `MULTI != RELIC_NONE` evaluates true.

It evaluates as false on some systems and since we cannot inherit relic's macros when compiling BLS,
we add an additional check to see if the function is defined (inheriting the macro's evaluation) and decide
accordingly if we make a call to it or not.

Tested on `Darwin Kernel Version 20.3.0: Thu Jan 21 00:07:06 PST 2021; root:xnu-7195.81.3~1/RELEASE_X86_64`